+++ Src/Library/Endpoint/EndpointDefinition.cs
@@ -43,8 +43,8 @@ public sealed class EndpointDefinition
     internal ServiceBoundEpProp[]? ServiceBoundEpProps;
     internal Action<RouteHandlerBuilder> InternalConfigAction;
     internal object RequestBinder;
-    internal object? PreProcessors;
-    internal object? PostProcessors;
+    internal List<object> PreProcessorList = new();
+    internal List<object> PostProcessorList = new();
     internal JsonSerializerContext? SerializerContext;
     internal bool ExecuteAsyncImplemented;
     internal ResponseCacheAttribute? ResponseCacheSettings { get; private set; }
@@ -277,6 +277,18 @@ public sealed class EndpointDefinition
         if (isScoped)
             ScopedValidator();
     }
+
+    /// <summary>
+    /// adds global pre-processors to this endpoint definition. these pre-processors are executed before the pre-processors configured at the endpoint level.
+    /// </summary>
+    /// <param name="preProcessors">the pre-processors to add</param>
+    public void PreProcessors(params IGlobalPreProcessor[] preProcessors) => PreProcessorList.AddRange(preProcessors.Where(p => !PreProcessorList.Any(x => x.GetType() == p.GetType())));
+
+    /// <summary>
+    /// adds global post-processors to this endpoint definition. these post-processors are executed before the post-processors configured at the endpoint level.
+    /// </summary>
+    /// <param name="postProcessors">the post-processors to add</param>
+    public void PostProcessors(params IGlobalPostProcessor[] postProcessors) => PostProcessorList.AddRange(postProcessors.Where(p => !PostProcessorList.Any(x => x.GetType() == p.GetType())));
 }
 
 /// <summary>