+++ Src/Library/changelog.md
@@ -34,6 +34,7 @@ app.UseFastEndpoints(c =>
 - code snippets added to [visual studio extension](https://marketplace.visualstudio.com/items?itemName=dj-nitehawk.FastEndpoints)
 - `FastEndpoints.ClientGen` package for c# and typescript client generation with `NSwag`
 - override default model binding logic by inheriting from `RequestBinder<TRequest>` class #189
+- global pre/post processor support [#info](https://discord.com/channels/933662816458645504/1011478168726683678)
 - `DontCatchExceptions()` method to enable custom exception handler middleware #186
 - `Validator<TValidafor>()` method for being explicit in the endpoint configuration #190
 - `IncludeAbstractValidators` startup flag for including validators inheriting `AbstractValidator<TRequest>` in auto registration #190